old_project/backend/requirements.txt
asgiref==3.8.1
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
gunicorn==20.1.0
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
Pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
whitenoise==6.6.0
django-redis==5.4.0
django-debug-toolbar==4.2.0
